Moved IVsBroswerService to Core
